+++ ow/__init__.py
@@ -1,1 +1,3 @@
-from .tile import Tile
+from .core.tile import Tile
+from .core.primitive import Primitive
+from .core.world import World