+++ ow/core/__init__.py
@@ -1,0 +1,3 @@
+from .tile import Tile
+from .primitive import Primitive
+from .world import World